Split CopyLink component from ShareInvite
